+++ src/h3sed/conf.py
@@ -26,7 +26,7 @@ import sys
 """Program title, version number and version date."""
 Name = "h3sed"
 Title = "Heroes3 Savegame Editor"
-Version = "2.5.dev1"
+Version = "2.5.dev2"
 VersionDate = "15.06.2024"
 
 Frozen = getattr(sys, "frozen", False)